+++ src/cli/helpers.js
@@ -1,5 +1,6 @@
 const fs = require('fs')
 const path = require('path')
+const execa = require('execa')
 const crypto = require('crypto')
 const { spawn, execSync } = require('child_process')
 const xxhash = require('xxhashjs')
@@ -22,16 +23,56 @@ const resolvePath = function (filepath) {
   return path.resolve(process.cwd(), filepath)
 }
 
-const executeCommand = function (subCommand, env) {
-  const subprocess = spawn(subCommand[0], subCommand.slice(1), {
-    stdio: 'inherit',
-    shell: true,
-    env: { ...process.env, ...env }
-  })
+const executeCommand = async function (subCommand, env) {
+  const signals = [
+    'SIGHUP', 'SIGQUIT', 'SIGILL', 'SIGTRAP', 'SIGABRT',
+    'SIGBUS', 'SIGFPE', 'SIGUSR1', 'SIGSEGV', 'SIGUSR2', 'SIGTERM'
+  ]
 
-  subprocess.on('close', (code) => {
-    if (code > 0) {
-      logger.error(`command [${subCommand.join(' ')}] failed (code: ${code})`)
+  logger.debug(`executing subcommand ${subCommand}`)
+
+  // handler for SIGINT
+  let subprocess
+  const sigintHandler = () => {
+    logger.debug('received SIGINT')
+    logger.debug('checking subprocess')
+    logger.debug(subprocess)
+
+    if (subprocess) {
+      logger.debug('sending SIGINT to subprocess')
+      subprocess.kill('SIGINT') // Send SIGINT to the subprocess
+    } else {
+      logger.debug('no subprocess to send SIGINT to')
+    }
+  }
+
+  const handleOtherSignal = (signal) => {
+    logger.debug(`received ${signal}`)
+  }
+
+  try {
+    subprocess = execa(subCommand[0], subCommand.slice(1), {
+      stdio: 'inherit',
+      env: { ...process.env, ...env }
+    })
+
+    process.on('SIGINT', sigintHandler)
+
+    signals.forEach(signal => {
+      process.on(signal, () => handleOtherSignal(signal))
+    })
+
+    // Wait for the subprocess to finish
+    const { exitCode } = await subprocess
+
+    if (exitCode !== 0) {
+      logger.debug(`received exitCode ${exitCode}`)
+      throw new Error(`Command failed with exit code ${exitCode}`)
+    }
+  } catch (error) {
+    if (error.signal !== 'SIGINT') {
+      logger.error(error.message)
+      logger.error(`command [${subCommand.join(' ')}] failed`)
       logger.error('')
       logger.error(`  try without dotenvx: [${subCommand.join(' ')}]`)
       logger.error('')
@@ -39,20 +80,12 @@ const executeCommand = function (subCommand, env) {
       logger.error(`<${REPORT_ISSUE_LINK}>`)
     }
 
-    process.exit(code)
-  })
-
-  subprocess.on('error', (err) => {
-    logger.error(err)
-    logger.error(`command [${subCommand.join(' ')}] failed`)
-    logger.error('')
-    logger.error(`  try without dotenvx: [${subCommand.join(' ')}]`)
-    logger.error('')
-    logger.error('if that succeeds, then dotenvx is the culprit. report issue:')
-    logger.error(`<${REPORT_ISSUE_LINK}>`)
-
-    process.exit(1)
-  })
+    // Exit with the error code from the subprocess, or 1 if unavailable
+    process.exit(error.exitCode || 1)
+  } finally {
+    // Clean up: Remove the SIGINT handler
+    process.removeListener('SIGINT', sigintHandler)
+  }
 }
 
 const pluralize = function (word, count) {